requirements.txt
Django==4.1.1
gunicorn
djangorestframework
django-environ==0.8.1
psycopg2